feat: force use of feature flag with compiler error
