+++ Cargo.toml
@@ -15,7 +15,7 @@ readme = "README.md"
 exclude = ["/tests"]
 
 [features]
-default=["gurobi11"]
+default=[]
 gurobi12 = ["grb-sys_12"]
 gurobi11 = ["grb-sys_10"]
 gurobi10 = ["grb-sys_10"]
@@ -23,6 +23,7 @@ gurobi9 = ["grb-sys_10"]
 
 
 [package.metadata.docs.rs]
+features = [ "gurobi12" ]
 rustdoc-args = ["--html-in-header", "doc/katex-header.html"]
 
 [dependencies]
@@ -30,7 +31,7 @@ fnv = "^1.0"
 anyhow = "^1.0"
 cstr-enum = "^1.0"
 grb-sys_10 = {package = "grb-sys2", version = "10.2.0", optional = true}
-grb-sys_12 = {package = "grb-sys2", path="../grb-sys2", version = "^12.0.0", optional = true}
+grb-sys_12 = {package = "grb-sys2", version = "^12.0.0", optional = true}
 grb-macro = "0.1.0"
 cfg-if = "1.0.0"
